refactor(sync): extract peer keystore + rebind guard into peers/peer_store.rs
